refactor: split QCSchema into multiple files
The qcschema.py file has grown quite substantially. In order to improve
maintainability and class extensions down the road, this refactoring
splits that single file into multiple files.
