+++ docs/apidocs/qiskit_nature.second_q.formats.qcschema.rst
@@ -1,0 +1,21 @@
+﻿QCSchema
+=======================================
+
+.. automodule:: qiskit_nature.second_q.formats.qcschema
+
+   .. rubric:: Classes
+
+   .. autosummary::
+
+      QCBasisSet
+      QCCenterData
+      QCECPPotential
+      QCElectronShell
+      QCError
+      QCModel
+      QCProperties
+      QCProvenance
+      QCSchema
+      QCSchemaInput
+      QCTopology
+      QCWavefunction
