+++ docs/apidocs/qiskit_nature.second_q.formats.rst
@@ -2,21 +2,3 @@
 ==============================
 
 .. automodule:: qiskit_nature.second_q.formats
-
-   .. rubric:: Classes
-
-   .. autosummary::
-
-      QCBasisSet
-      QCCenterData
-      QCECPPotential
-      QCElectronShell
-      QCError
-      QCModel
-      QCProperties
-      QCProvenance
-      QCSchema
-      QCSchemaInput
-      QCTopology
-      QCWavefunction
-
